Refactor sensor and radar struct (#17)
* Refactor SPI device functionality from u16 to u8

* rustfmt

* Add interrupt waiting after radar measurement

* Implement Debug trait for RadarData struct

* Refactor radar and sensor classes for better management

This commit refactors the radar and sensor classes for better error tracking and state management. It adds and updates functions for calibrating, measuring, and reading sensor data. Furthermore, this update gives the ability to check radar connection and sensor status. The TransitionError struct has also been enhanced with the Debug trait for improved error handling and tracking.

* Update Cargo.lock